renamed contextualizer and providedservices
